integrated backend and frontend
