--- apps/backend/src/routes/uploadRoutes.js
@@ -1,9 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const { uploadFiles } = require('../controllers/uploadController');
-
-// POST route for file uploads
-router.post('/', uploadFiles);
-
-module.exports = router;
-
