+++ apps/backend/webpack.config.js
@@ -5,10 +5,6 @@ module.exports = {
   output: {
     path: join(__dirname, '../../dist/apps/backend'),
   },
- 
-  resolve: {
-    extensions: ['.ts', '.js'],
-  },
   plugins: [
     new NxAppWebpackPlugin({
       target: 'node',
@@ -22,4 +18,3 @@ module.exports = {
     }),
   ],
 }
-
