+++ package.json
@@ -18,6 +18,7 @@
     "@nrwl/web": "^19.8.4",
     "axios": "^1.6.0",
     "express": "^4.18.1",
+    "multer": "^1.4.5-lts.1",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-router-dom": "6.11.2"
